Move Type enum to Item from Parser
